requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
diff-match-patch==20230430
Django==5.0.6
django-cors-headers==4.4.0
django-filter==24.2
django-import-export==4.1.0
django-phonenumber-field==8.0.0
django-rest-knox==4.2.0
djangorestframework==3.15.2
gunicorn==22.0.0
idna==3.7
packaging==24.1
phonenumbers==8.13.40
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
requests==2.32.3
sqlparse==0.5.0
tablib==3.5.0
typing_extensions==4.12.2
urllib3==2.2.2